# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movie/Show Watchlist API",
    "version": "1.0.0",
    "description": "A REST API for managing a personal movie and show watchlist. Write operations (POST, PUT, DELETE) require a logged-in session via GitHub OAuth. Because auth is session/cookie based, protected routes must be tested in a browser after logging in at /auth/github — the Swagger \"Try it out\" button cannot hold the login session, so protected calls will return 401 there. Read operations (GET) are public."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "githubSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie set after logging in via GitHub OAuth at /auth/github. Established in the browser, not through Swagger UI."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Auto-generated MongoDB id",
            "readOnly": true
          },
          "githubId": {
            "type": "string",
            "example": "1234567"
          },
          "username": {
            "type": "string",
            "example": "octocat"
          },
          "email": {
            "type": "string",
            "example": "octocat@github.com"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Title": {
        "type": "object",
        "required": [
          "title",
          "type",
          "releaseYear",
          "status",
          "genreId"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Auto-generated MongoDB id",
            "readOnly": true
          },
          "title": {
            "type": "string",
            "example": "Inception"
          },
          "type": {
            "type": "string",
            "enum": [
              "movie",
              "show"
            ],
            "example": "movie"
          },
          "releaseYear": {
            "type": "integer",
            "example": 2010
          },
          "runtime": {
            "type": "integer",
            "description": "Runtime in minutes",
            "example": 148
          },
          "status": {
            "type": "string",
            "enum": [
              "watchlist",
              "watching",
              "finished"
            ],
            "example": "finished"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "example": 9
          },
          "dateWatched": {
            "type": "string",
            "format": "date",
            "example": "2026-01-15"
          },
          "genreId": {
            "type": "string",
            "description": "ObjectId referencing a genre",
            "example": "65a1f2c3d4e5f6a7b8c9d0e1"
          }
        }
      },
      "Genre": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Auto-generated MongoDB id",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "example": "Science Fiction"
          },
          "description": {
            "type": "string",
            "example": "Films and shows set in speculative, futuristic, or scientific worlds"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Resource not found"
          }
        }
      }
    }
  },
  "paths": {
    "/titles": {
      "get": {
        "summary": "Get all titles",
        "tags": [
          "Titles"
        ],
        "responses": {
          "200": {
            "description": "A list of all titles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Title"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new title",
        "tags": [
          "Titles"
        ],
        "security": [
          {
            "githubSession": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Title"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created title",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Title"
                }
              }
            }
          },
          "400": {
            "description": "Validation error (missing/invalid fields)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/titles/{id}": {
      "get": {
        "summary": "Get a single title by id",
        "tags": [
          "Titles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The MongoDB ObjectId of the title"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested title",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Title"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Title not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an existing title",
        "tags": [
          "Titles"
        ],
        "security": [
          {
            "githubSession": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The MongoDB ObjectId of the title"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Title"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated title",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Title"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid id format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Title not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a title",
        "tags": [
          "Titles"
        ],
        "security": [
          {
            "githubSession": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The MongoDB ObjectId of the title"
          }
        ],
        "responses": {
          "200": {
            "description": "Title deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Title deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid id format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Title not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/genres": {
      "get": {
        "summary": "Get all genres",
        "tags": [
          "Genres"
        ],
        "responses": {
          "200": {
            "description": "A list of all genres",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Genre"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new genre",
        "tags": [
          "Genres"
        ],
        "security": [
          {
            "githubSession": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Genre"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created genre",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Genre"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or duplicate genre name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/genres/{id}": {
      "get": {
        "summary": "Get a single genre by id",
        "tags": [
          "Genres"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The MongoDB ObjectId of the genre"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested genre",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Genre"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Genre not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an existing genre",
        "tags": [
          "Genres"
        ],
        "security": [
          {
            "githubSession": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The MongoDB ObjectId of the genre"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Genre"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated genre",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Genre"
                }
              }
            }
          },
          "400": {
            "description": "Validation error, duplicate name, or invalid id format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Genre not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a genre",
        "tags": [
          "Genres"
        ],
        "security": [
          {
            "githubSession": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The MongoDB ObjectId of the genre"
          }
        ],
        "responses": {
          "200": {
            "description": "Genre deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Genre deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid id format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Genre not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/github": {
      "get": {
        "summary": "Start the GitHub OAuth login flow",
        "tags": [
          "Auth"
        ],
        "description": "Redirects the browser to GitHub to authorize the app. Open this in a browser, not the Swagger \"Try it out\" button.",
        "responses": {
          "302": {
            "description": "Redirect to GitHub for authorization"
          }
        }
      }
    },
    "/auth/github/callback": {
      "get": {
        "summary": "GitHub OAuth callback",
        "tags": [
          "Auth"
        ],
        "description": "GitHub redirects here after authorization. Creates the session and redirects to the API docs on success.",
        "responses": {
          "302": {
            "description": "Redirect to /api-docs on success or /auth/status on failure"
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "summary": "Log out and destroy the session",
        "tags": [
          "Auth"
        ],
        "responses": {
          "302": {
            "description": "Redirect to /auth/status after logging out"
          }
        }
      }
    },
    "/auth/status": {
      "get": {
        "summary": "Check current authentication state",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Whether a user is logged in, and their info if so",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "loggedIn": {
                      "type": "boolean",
                      "example": true
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Titles",
      "description": "CRUD operations for movies and shows on the watchlist"
    },
    {
      "name": "Genres",
      "description": "Lookup collection of genres referenced by titles"
    },
    {
      "name": "Auth",
      "description": "GitHub OAuth authentication"
    }
  ]
}